@@ -13,14 +13,15 @@
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<java.version>1.8</java.version>
+		<java.version>11</java.version>
 
 		<hikari.version>3.4.5</hikari.version>
 		<spring.version>5.3.22</spring.version>
 		<log4jframework.version>2.18.0</log4jframework.version>
 		<mysql.driver.version>8.0.29</mysql.driver.version>
-		<junit.version>4.13.2</junit.version>
-		<maven.compiler.plugin.version>3.6.1</maven.compiler.plugin.version>
+		<junit.version>5.9.0</junit.version>
+		<maven.compiler.plugin.version>3.10.1</maven.compiler.plugin.version>
+		<maven.surefire.plugin.version>2.22.2</maven.surefire.plugin.version>
 	</properties>
 
 
@@ -39,8 +40,14 @@
 					</configuration>
 				</plugin>
 
-			</plugins>
+				<plugin>
+					<groupId>org.apache.maven.plugins</groupId>
+					<artifactId>maven-surefire-plugin</artifactId>
+					<version>${maven.surefire.plugin.version}</version>
+				</plugin>
 
+
+			</plugins>
 		</pluginManagement>
 	</build>
 
@@ -95,8 +102,8 @@
 		</dependency>
 
 		<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
+			<groupId>org.junit.jupiter</groupId>
+			<artifactId>junit-jupiter-engine</artifactId>
 			<version>${junit.version}</version>
 			<scope>test</scope>
 		</dependency>
